Add tiken verification

+++ pages/api/products/[id].js
@@ -4,8 +4,11 @@ import Product from '../../../models/Product';
 export default async function handler(req, res) {
   const {
     method,
+    cookies,
     query: { id },
   } = req;
+  const { token } = cookies;
+
   await dbConnect();
   if (method === 'GET') {
     try {
@@ -24,6 +27,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === 'PUT') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('You are not authenticated');
+    }
     try {
       const product = await Product.create(req.body);
       res.status(201).json(product);
@@ -32,6 +38,9 @@ export default async function handler(req, res) {
     }
   }
   if (method === 'DELETE') {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json('You are not authenticated');
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json('Product deleted successfully!');
